requirements.txt
astroid==3.3.10
certifi==2025.4.26
charset-normalizer==3.4.2
dill==0.4.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pylint==3.3.7
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
urllib3==2.4.0
